+++ Math/math_isnan.cpp
@@ -1,15 +1,15 @@
 #include "math.hpp"
 
-int math_isnan(double number)
+int32_t math_isnan(double number)
 {
     union
     {
         double double_value;
-        unsigned long long integer_value;
+        uint64_t integer_value;
     } converter;
 
-    unsigned long long exponent_bits;
-    unsigned long long mantissa_bits;
+    uint64_t exponent_bits;
+    uint64_t mantissa_bits;
 
     converter.double_value = number;
     exponent_bits = converter.integer_value & 0x7ff0000000000000ULL;
