+++ Math/math_sqrt.cpp
@@ -2,16 +2,16 @@
 #include "../Errno/errno.hpp"
 #include <cfloat>
 
-static int math_is_infinite_internal(double number)
+static int32_t math_is_infinite_internal(double number)
 {
     union
     {
-        double              double_value;
-        unsigned long long  integer_value;
+        double      double_value;
+        uint64_t    integer_value;
     }   converter;
 
-    unsigned long long exponent_bits;
-    unsigned long long mantissa_bits;
+    uint64_t exponent_bits;
+    uint64_t mantissa_bits;
 
     converter.double_value = number;
     exponent_bits = converter.integer_value & 0x7ff0000000000000ULL;
@@ -27,8 +27,8 @@ double math_sqrt(double number)
     double next_guess;
     double difference;
     double tolerance;
-    int    iteration_count;
-    int    max_iterations;
+    int32_t    iteration_count;
+    int32_t    max_iterations;
 
     if (math_isnan(number))
     {
