+++ System_utils/System_utils_sysinfo.cpp
@@ -3,18 +3,18 @@
 #include "../Compatebility/compatebility_internal.hpp"
 #include "../Errno/errno.hpp"
 
-unsigned int    su_get_cpu_count(void)
+uint32_t    su_get_cpu_count(void)
 {
-    unsigned int result;
+    uint32_t result;
 
     result = cmp_get_cpu_count();
     return (result);
 }
 
-unsigned long long su_get_total_memory(void)
+uint64_t su_get_total_memory(void)
 {
     uint64_t result;
-    int error_code;
+    int32_t error_code;
 
     result = 0;
     error_code = cmp_get_total_memory(&result);
